ZJIT: Allow higher profile num (#14698)
When we investigate guard failure issues, we sometimes need to use
profile num around 100k (e.g. `lobsters` in ruby-bench).

This change is to allow that.

+++ zjit/src/profile.rs
@@ -278,7 +278,7 @@ pub struct IseqProfile {
     opnd_types: Vec<Vec<TypeDistribution>>,
 
     /// Number of profiled executions for each YARV instruction, indexed by the instruction index
-    num_profiles: Vec<u8>,
+    num_profiles: Vec<u32>,
 }
 
 impl IseqProfile {
